Move environment variable related functions to own file
